Make the SpinnerView centered in its parent view
Although the spinnerView is 'more centered' if it's positioned centered in the current controller's navigationController's view, it means the spinnerView is tied to the wrong view when we transition between controllers. It'd be better to figure out a way to manually position it, while still keeping it a subview of the current view controller's view.