Tear out TestFlightSDK
Apple acquired and shut down the existing TestFlightSDK, including the
CocoaPod for it (see
http://9to5mac.com/2015/01/26/apple-legacy-testflight-shutdown/ for
instance). I can’t build the app with this pod required, and it looks
to me like moving to the new TestFlight
(https://developer.apple.com/testflight/) is a big revision. This is
intended to be a minimal “just get things building” change.